requirements.txt
numpy==1.13.3
Pillow==4.2.1
PyOpenGL==3.1.1a1
PyOpenGL-accelerate==3.1.1a1
ftputil==3.3.1
pygame==1.9.3
cython==0.26.1